requirements.txt
streamlit
numpy
opencv-python-headless
tensorflow-cpu
scikit-learn
matplotlib
